+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Add option to show overlay through terrain.
+
 ### 1.0.6
 - Update translations.